Refactor routes to use controllers
