+++ back/src/routes/betRouter.ts
@@ -1,0 +1,5 @@
+// src/routes/betRouter.ts
+
+import express, { Request, Response } from 'express';
+
+export const betRouter = express.Router();
